Refactor logic between controller and domain class
Refs hitobito/hitobito#342
