--- app/models/tag.rb
@@ -1,9 +1,0 @@
-# encoding: utf-8
-
-#  Copyright (c) 2019, hitobito AG. This file is part of
-#  hitobito and licensed under the Affero General Public License version 3
-#  or later. See the COPYING file at the top-level directory or at
-#  https://github.com/hitobito/hitobito.
-
-class Tag < ActsAsTaggableOn::Tag
-end
